Add version number when publishing metrics.
Makes it easier to query server side by version without having to
write joins. Note this is the version number of the publisher,
not the tracker. E.g. for emulator it will still need to log the
emulator version to track that version separate from the version
of Android Studio it is used with.

Change-Id: I2b37be23b6d3a2ee57e3c6c355b5c6c7c36dbf6d