Usage Tracker: add product version to logs
Right now we only track the version of Android Studio
doing the upload of the metrics. This can cause
really weird behavior if the user is running
two different versions of Android Studio Side-by-Side.
As the tracking and publishing are separated we can end up
 with:
-metrics reported for newer version
-metrics reported for older version

So instead we now track for each metric the version it was logged
as well as the version it was published with.

We can then process this on the server side.
Bug: http://b.android.com/232515
Test: Testcase added

Change-Id: I9a59af65b518166ea0c05e8b961af40325eeb8dc
(cherry picked from commit ea6f7f10ccb321012d4da61cbed41dc85e3d6e3a)