UsageTracker stops logging after error
If an exception is raised when switching track file or writing out
data, JournalingUsageTracker is not closed and set in Broken state.
Such error is typically a result of running out if disk space or
file permission issues and not much can be done to overcome the
problem.

Issue:
Currently, if such error occurs, usage tracker may be left in a
undefined state, where some of its fields are initialized, but others
are not. Any future log events will result in another exception.

Fix:
If any error occurs JournalingUsageTracker is now in a predictable
Broken state. IOExceptions are ignored when opening/writing to the
log. These are not critical and should not show up as such.

Other fixes/changes:
* UsageTracker.initialize is now responsible for closing previous
  instances.
* UsageTracker.updateSettingsAndTracker doesn't throw anymore if
  there is a problem with initialization of a new tracker.

Test: None
Fixes: 77968950

Change-Id: I4a7dbe23b55cc95feef8789e952020bdedb53282