Spool file channel should specify DSYNC
Currently we're manually flushing the output stream, then try to force
the channel to flush (on content changes only).

On Windows, this results in calling FlushFileBuffers, which can be
inefficient if used after each and every write operation. With this
change, we're now passing FILE_FLAG_WRITE_THROUGH to CreateFile, and on
Linux we're using O_DSYNC, which should be slightly better.

Bug: N/A
Change-Id: Id02842d6e99aa6a8e0ebaaa663e9d90314cb1a91