studio stats: clean up AnalyticsSettings API
AnalyticsSettings API used to be instance based, this was exposing an
implementation detail of the GSON object being serialized.

This caused issues as people kept cached instances around.

This change makes all public API static on AnalyticsSettings
and creats a separate AnalyticsSettingsData object that
is only used for serialization and testing.

Test: updated
Bug: 110560339
Change-Id: I2ee06d1ad28f7649651b74a96f54298a069a3502