report analytics use before init in debug builds
We had a few attempts to use analytics before they were initialized.
This adds a check when running with idea.is.internal to warn
about this case.

As this check only throws when idea.is.internal, we rely on the default
behavior of not initialized returning false values and noop actions.

This ensures that any issues we missed will not cause crashes in production

Fixes: 112698125
Test: added

Change-Id: I60e4b9ec0740cb1c286b04b64e2bc951020f579e