Flush events on UsageTracker.deinitiliaze()
Previously pending events might have been dropped when the
Usage tracker was de-initialized at the end of a Gradle invocation.

Makes UsageTrackerWriter implement Flushable, so that close can keep the
same semantic, but flush will ensure that events logged before flush was
called are not lost.

Instead of creating a new Runnable for each LogEvent that is posted,
keep pending events in a Queue and schedule a Runnable to clear that
queue.

Change-Id: I17347c876d1315c59fd3ed97f25c06c5e4cb0ab4
Fixes: 123403180
Test: Added cases to JournalingUsageTrackerTest, Manually checked AGP behavior.