Update clearcut proto to support instant apps CLI
We're adding some logging to the IA CLI and need a Maven build target
for the logging protos. Since Studio already uses this version, the most
expedient solution is to add the required fields here.

Note that the auto-generation script mentioned at the top of the proto
file no longer exists. I did explore the possibility of writing my own
script (see cl/235303885), but the result adds 1900 new lines to the
proto definition. Auditing all those lines and redacting them, not to
mention shipping those extra enum values in the JAR, seems like an
inferior option to just adding the fields we need when we need them
(which has evidently been very rare in the past).

Test: Existing Studio tests + manual test building IA CLI
Bug: 124066490
Change-Id: Ic8a2ff1cb00fb4a7f0de422eff9c01d15c70b3ff