Make AnalyticsInspector tool ask for path if not provided
It may be useful for Analytics inspector to ask for a file path when
not provided as a parameter instead of just exit.
e.g. Makes it possible to run from ide and not save path in run
configuration.
In case of providing path as command line parameter nothing changes.

Test: manually verified
Bug: N/A

Change-Id: I945be8b2c580fa8d3d4daa67eb2528acfdb2ce4d
Signed-off-by: Maksim Lazeba <mlazeba@google.com>