Add ENABLE_CMAKE_BUILD_COHABITATION flag proto
This proto is required when adding a new command-line flag
to Android Gradle Plugin. It's used to track metrics about
whether users turn to flag off or on.

If there is a value in com.android.build.gradle.options.BooleanOption
then there must be a corresponding entry in this proto.

Bug: 130521279
Test: NA
Change-Id: I9ddab260225a598690595376925a069b02bf357d