Add missing task enums to analytics_enums.proto
Previously, AnalyticsUtilTest was ignoring abstract tasks when
checking that all tasks have corresponding enums.

This CL adds all abstract tasks to analytics_enums.proto. This is a bit
messy because some of these abstract tasks will never show up in the
analytics data, but it seems safer to include all of them (as opposed to
ignoring the subset which don't show up in the analytics data) in case
some of them are edited in the future and end up in the analytics data.

Bug: NA
Test: AnalyticsUtilTest
Change-Id: I45f15775bae9bd635be0c03ca505fcddd4548c77