Explcitly list all env vars and sys properties in analytics
This is needed so that we can fetch all values from AGP. AGP
implementation of Environment is retained by analytics lib
across builds, and because Gradle APIs are used to obtain values,
this caused leaks such as b/160330055. With this change, all
values can be obtained and there is no need to retain any Gradle
state.

Bug: 160330055
Test: existing
Change-Id: I253b9e1bdb715beec658952b9617f36dd738a27a