Clean up confusion between androidHome and SDK dir.
Android Home served as the name of the preferences
folder (generally ~/.android) and could be overriden
via the property ANDROID_SDK_HOME

The SDK location was overrideable via the property
ANDROID_HOME.

This led to confusion in the code, naming of variable
that was not obvious (often containing Sdk to match
the property name instead of the content).

This commit cleans this up, introduce a new
ANDROID_PREFS_ROOT to replace ANDROID_SDK_HOME (matching
a new ANDROID_SDK_ROOT introduced a while back to replace
ANDROID_HOME).

Bug: N/A
Test: existing tests have been updated.
Change-Id: Idc83e7d77a07d0379994aebc636fb5967604f1a5