Update the stubbed GraphicsDevice to not assume headless mode
AndroidStudioUsageTrackerTest.testGetMachineDetails fails when running
from the IDE on Windows:

java.lang.AssertionError
	at org.junit.Assert.fail(Assert.java:86)
	at org.junit.Assert.fail(Assert.java:95)
	at com.android.tools.analytics.stubs.StubGraphicsConfiguration.getDevice(StubGraphicsConfiguration.kt:34)
	at com.intellij.ui.scale.JBUIScale.sysScale(JBUIScale.java:278)
	at com.android.tools.idea.stats.AndroidStudioUsageTracker.getDisplayDetails(AndroidStudioUsageTracker.kt:152)
	at com.android.tools.idea.stats.AndroidStudioUsageTracker.getMachineDetails(AndroidStudioUsageTracker.kt:172)
	at com.android.tools.idea.stats.AndroidStudioUsageTrackerTest.testGetMachineDetails(AndroidStudioUsageTrackerTest.kt:104)

This is because JBUIScale.sysScale determines HiDPI status based on
the running JRE and the java.awt.headless property. The headless
property was only set during IdeaTestSuiteBase initialization, as
part of the early TestApplication setup, which explains why the test
passes CI and only fails when running by itself.

We're trying to clean up the early TestApplication initialization,
which means StubGraphicsDevice has to work no matter what the headless
status is.

Bug: 193111860
Test: AndroidStudioUsageTrackerTest
Change-Id: Idaea660929579810ca88552d736956ea3261597c