Expose whether the UsageTracker is initialized
This is neccessary for lint to make sure it initializes
the usage tracker if it hasn't already been done. When
running in process with AGP, AGP will have initialized it,
but when running as a separate command, it has not.
Using just AnalyticsSettings.initialized as a proxy didn't
work well since for example during unit testing the
analytics tests will initialize the settings but close
the usage tracker when done.

Bug: 194525628
Test: N/A (just making a getter public; functionality tested in lint)
Change-Id: Ie5cf067c7918bc27764d95790075fa9fe0b3f157