Stop scheduling unnecessary flush operations
Each UsageTracker's lgo call will schedule a flush on a background
thread. Flush has a synchronized block - if a single flush
takes a long time, all other threads will wait for it to finish.

This can result in thousands of background threads waiting to
acquire the lock.

All these threads are unnecessary to start if flush is already
happening.

The fix contains two improvements:

* New logs will only schedule flush operation if no other call scheduled
  it yet and flush is not currently running

* tryFlush skips flushing if one is already happening

Both tryFlush and flush will schedule another flush at the end if there
is any item pending to be flushed. This avoids concurrency problem
where it would be possible to add item to the pending queue with no
pending flush, keeping the event way too long in the queue.

flush() is still blocking operation as it is required to properly
handle pending events on app close.

Fixes: 295909822
Test: Manual, triggered high use of usage tracker while monitoring
  number of thread from the thread pool
Change-Id: I4df060b2774ffae77fc394252eb0e9b674e01f65