Set connection timeout on analytics publisher test
testBadServer is hanging and being terminated by the
timeout rule. This can interfere with other tests because
the gate is not released. Use connection timeouts instead
so that other tests are not affected.

This does not address the issue of why testBadServer is timing
out but will prevent testBadServer failures from affecting
other tests.

Bug: 295070308
Test: Self
Change-Id: Ice4c5398eb80a016060ff7133d73a07a40a6d520