Avoid uploading too-large files for truncated fields
Currently, if a field is larger than 250KB, it will be truncated and its
full contents attached as a file to the crash report. If it is larger
than 1.2MB (the Crash file size limit), Crash will discard _all_ file
attachments, so attempting to attach the full contents as a file is
counterproductive in this case.

Change-Id: I0797a22699ad656cca4a6bb3ed33abdb302efcac