Add SIMPLIFIED_MERGED_MANIFESTS_PRODUCER to proto
The issue is that the MERGED_MANIFEST and MERGED_MANIFESTS
are produced by the same ProcessApplicationManifest task.
That's completely wrong because we package the MERGED_MANIFEST
inside the MERGED_MANIFESTS. It works for `application` because
the ProcessMultiApkApplicationManifest is resetting itself
as the initial provider of the MERGED_MANIFESTS so for app
you can have the right chaining of ProcessApplicationManifest ->
manifestUpdater -> ProcessMultiApkApplicationManifest.

However, when ProcessMultiApkApplicationManifest is not
running then it's broken because the same task provides
both MERGED_MANIFEST and MERGED_MANIFESTS not allowing
the MERGED_MANIFEST to be updated.

SimplifiedMergedManifestsProducer was added for dynamic feature
projects so user can update manifest and introduced  task will
consume it.

Bug: 321141574
Test: NA
Change-Id: I6a1d0ad2afdc9a09bef6aead1a61381b846c95d2