Remove failing steps from AnalyticsPublisherTest
The last few steps of this test fail in presubmit
very often but pass when debugging locally. Remove
the extra steps as the test has already verified
that the upload succeeds after the server recovers.

It would be better to solve the problem but it does
not seem worth the cost.

Bug: 330783405
Test: Self
Change-Id: I32b4e013c53d9ae4a78eeeac33fed82f57e30094