Fix Windows ARM64 detection
https://learn.microsoft.com/en-us/windows/arm/apps-on-arm-x86-emulation#detecting-emulation
"For compatibility reasons, the API GetNativeSystemInfo will also
return emulated processor details when run from an app under
emulation."

We still use an x86_64 JNA native library, so our only option is the
IsWow64Process2 API. It's not yet exposed via JNA, but we can call
it directly. We manually verified that the native architecture result
is correct and returns -21916 or 0xAA64 for ARM64.

Bug: 365768259
Test: manual
Change-Id: I488db387c6d8758cbc52f18a8c00423fb465146a